connect-editor-with-selected-question/updated logic

+++ admin/lib/presentation/widgets/customization_panel/intro/intro_common_customization_tab.dart
@@ -1,9 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:survey_admin/presentation/app/localization/localizations.dart';
-import 'package:survey_admin/presentation/utils/app_fonts.dart';
-import 'package:survey_admin/presentation/utils/colors.dart';
-import 'package:survey_admin/presentation/utils/constants/app_dimensions.dart';
 import 'package:survey_admin/presentation/widgets/customization_items/color_customization_item.dart';
 import 'package:survey_admin/presentation/widgets/customization_items/customization_items_container.dart';
 import 'package:survey_admin/presentation/widgets/customization_items/radius_customization_item.dart';
@@ -121,7 +118,7 @@ class IntroCommonCustomizationTab extends CustomizationTab {
               },
             ),
             RadiusCustomizationItem(
-              initialValue: initialTheme.mainButtonRadius.toInt(),
+              initialValue: initialTheme.mainButtonRadius,
               onRadiusChanged: onButtonRadiusChanged,
             ),
           ],
